I added UserDefault functionality, Tom expressed that he wasn't into having an unfilter, but rather just manipulate the text after storing it without the filter. Thoughts?
