+++ BYT/Views/FoaasSettingsMenuView.swift
@@ -26,6 +26,7 @@ class FoaasSettingsMenuView: UIView, UIScrollViewDelegate {
     
     override init(frame: CGRect) {
         super.init(frame: frame)
+        self.updateSwitch()
         if let view = Bundle.main.loadNibNamed("FoaasSettingsMenuView", owner: self, options: nil)?.first as? UIView {
             //add a view to the subview
             self.addSubview(view)
@@ -118,6 +119,14 @@ class FoaasSettingsMenuView: UIView, UIScrollViewDelegate {
         return view
     }()
     
+    
+    //MARK: - Helper Function
+    func updateSwitch () {
+        
+        //I'm not fully sure as why I need update this here instead of App Delegate, but whenever I put this in appdelegate it did not seem to update the languageFilterToggle early enough? If anyone knows why this is the case that would be awesome, I suspect it has something to do with the timing of when a xib gets loaded.
+        LanguageFilterManager.shared.checkCurrentToggle()
+        self.profanitySwitch.setOn(languageFilterToggle, animated: false)
+    }
 }
